Struct holochain_sodium::secbuf::SecBuf

pub struct SecBuf { /* fields omitted */ }

A SecBuf is a memory buffer for use with libsodium functions. It can be backed by insecure (raw) memory for things like public keys, or secure (mlocked / mprotected) memory for things like private keys.

Methods

impl SecBuf

pub fn with_insecure(s: usize) -> Self

create a new SecBuf backed by insecure memory (for things like public keys)

pub fn with_secure(s: usize) -> Self

create a new SecBuf backed by secure memory (for things like private keys) warning: funky sizes may result in mis-alignment

pub fn with_insecure_from_string(s: String) -> Self

pub fn protect_state(&self) -> ProtectState

what is the current memory protection state of this SecBuf?

pub fn len(&self) -> usize

should be able to get size without messing with mem protection

pub fn readable(&mut self)

make this SecBuf readable

pub fn writable(&mut self)

make this SecBuf writable

pub fn noaccess(&mut self)

secure this SecBuf against reading or writing

pub fn read_lock(&mut self) -> Locker

make this SecBuf readable, and return a locker object that will secure this SecBuf automatically when it goes out of scope.

pub fn write_lock(&mut self) -> Locker

make this SecBuf writable, and return a locker object that will secure this SecBuf automatically when it goes out of scope.

pub fn write(&mut self, offset: usize, data: &[u8]) -> Result<(), SodiumError>

helper for writing data to our internal buffer

impl SecBuf

pub fn randomize(&mut self)

randomize the provided SecBuf

impl SecBuf

pub fn zero(&mut self)

Zero all memory

pub fn increment(&mut self)

Increments all memory by 1

pub fn compare(&mut self, b: &mut SecBuf) -> i32

Compares two SecBuf Return : | if a > b; return 1 | if a < b; return -1 | if a == b; return 0

pub fn from_array(&mut self, data: &[u8]) -> Result<(), SodiumError>

Load the u8 into the SecBuf